menu item bounding box improvement
this is a lot. but it works. went into `pict3d` cod and the current bad
bounding box behaviour for partial cylinders is unfixable w/ my tired
brain. for future self: the `get-cylinder-shape-bbox` function, and
probably in turn `transformed-disk-flrect3` as well, need to (if
possible?) take into account the `max-angle` of the thing. For the sake
of bounding boxes, everything is considered a full circle right now,
which leads to some odd looking bounding boxes.